+++ test/p2p/host/autorelay/relay_finder_test.mocks.dart
@@ -3,6 +3,7 @@
 // Do not manually edit this file.
 
 // ignore_for_file: no_leading_underscores_for_library_prefixes
+
 import 'dart:async' as _i14;
 
 import 'package:ipfs_libp2p/config/config.dart' as _i20;